+++ src/pc/network/packets/packet_object.c
@@ -263,13 +263,13 @@ bool network_set_sync_id(struct Object* o) {
 
 static void packet_write_object_header(struct Packet* p, struct Object* o) {
     struct SyncObject* so = &gSyncObjects[o->oSyncID];
-    u16 behaviorId = get_id_from_behavior(o->behavior);
+    u32 behaviorId = get_id_from_behavior(o->behavior);
 
     packet_write(p, &gNetworkPlayerLocal->globalIndex, sizeof(u8));
     packet_write(p, &o->oSyncID, sizeof(u32));
     packet_write(p, &so->txEventId, sizeof(u16));
     packet_write(p, &so->randomSeed, sizeof(u16));
-    packet_write(p, &behaviorId, sizeof(u16));
+    packet_write(p, &behaviorId, sizeof(u32));
 }
 
 static bool allowable_behavior_change(struct SyncObject* so, BehaviorScript* behavior) {
@@ -333,8 +333,8 @@ static struct SyncObject* packet_read_object_header(struct Packet* p, u8* fromLo
     packet_read(p, &so->randomSeed, sizeof(u16));
 
     // make sure the behaviors match
-    u16 behaviorId;
-    packet_read(p, &behaviorId, sizeof(u16));
+    u32 behaviorId;
+    packet_read(p, &behaviorId, sizeof(u32));
 
     BehaviorScript* behavior = (BehaviorScript*)get_behavior_from_id(behaviorId);
     if (behavior == NULL) {
