+++ src/pc/network/packets/packet_spawn_objects.c
@@ -17,7 +17,7 @@
 struct SpawnObjectData {
     u8 parentId;
     u32 model;
-    u16 behaviorId;
+    u32 behaviorId;
     s16 activeFlags;
     s32 rawData[80];
     u8 globalPlayerIndex;
@@ -73,13 +73,13 @@ void network_send_spawn_objects_to(u8 sendToLocalIndex, struct Object* objects[]
         struct Object* o = objects[i];
         u32 model = models[i];
         u8 parentId = generate_parent_id(objects, i, true);
-        u16 behaviorId = get_id_from_behavior(o->behavior);
+        u32 behaviorId = get_id_from_behavior(o->behavior);
         u8 extendedModelId = (o->oSyncID != 0 && gSyncObjects[o->oSyncID].o == o)
                            ? gSyncObjects[o->oSyncID].extendedModelId
                            : 0xFF;
         packet_write(&p, &parentId, sizeof(u8));
         packet_write(&p, &model, sizeof(u32));
-        packet_write(&p, &behaviorId, sizeof(u16));
+        packet_write(&p, &behaviorId, sizeof(u32));
         packet_write(&p, &o->activeFlags, sizeof(s16));
         packet_write(&p, o->rawData.asU32, sizeof(s32) * 80);
         packet_write(&p, &o->header.gfx.scale[0], sizeof(f32));
@@ -115,7 +115,7 @@ void network_receive_spawn_objects(struct Packet* p) {
         Vec3f scale = { 0 };
         packet_read(p, &data.parentId, sizeof(u8));
         packet_read(p, &data.model, sizeof(u32));
-        packet_read(p, &data.behaviorId, sizeof(u16));
+        packet_read(p, &data.behaviorId, sizeof(u32));
         packet_read(p, &data.activeFlags, sizeof(s16));
         packet_read(p, &data.rawData, sizeof(s32) * 80);
         packet_read(p, &scale[0], sizeof(f32));
